Multiple fixes: safety casts, formatting, and OpenAL
Apply a collection of safety, correctness and build fixes across the codebase. Key changes include: add explicit bounds checks and static_casts to avoid signed/unsigned and overflow issues (botlib, client, audio, json, etc.); replace unsafe sprintf usage with Com_sprintf for bounded formatting; refactor the VM DROP macro into a VM_Drop helper for safer variadic error reporting; add pkg-config fallback for OpenAL in CI and include OpenAL compatibility defines; enable conditional depth-fade handling in Vulkan fragment shader template; minor API cleanups in the OpenAL audio world/backend and small binary updates to SPIR-V shader data. Also include a few behavioral fixes (input run flag comparison, snapshot/areamask size checks, makefile command markers). These changes improve robustness and cross-platform build reliability.

+++ code/client/audio/legacy/snd_dma.cpp
@@ -1280,7 +1280,7 @@ static void S_Update_( int msec ) {
 		& ~(dma.submission_chunk-1);
 
 	// never mix more than the complete buffer
-	if ( endtime - s_paintedtime > dma.fullsamples ) {
+	if ( endtime - s_paintedtime > static_cast<unsigned>( dma.fullsamples ) ) {
 		endtime = s_paintedtime + dma.fullsamples;
 	}
