+++ code/win32/win_qvk.cpp
@@ -139,7 +139,7 @@ qboolean QVK_Init( void )
 		};
 		int i;
 
-		for ( i = 0; i < ARRAY_LEN( dllnames ); i++ )
+		for ( i = 0; i < static_cast<int>( ARRAY_LEN( dllnames ) ); i++ )
 		{
 			glw_state.VulkanLib = load_vulkan_library( dllnames[i] );
